+++ jeepney/io/blocking.py
@@ -1,5 +1,6 @@
 """Synchronous IO wrappers around jeepney
 """
+import array
 from collections import deque
 from errno import ECONNRESET
 import functools
@@ -9,14 +10,22 @@ from selectors import DefaultSelector, EVENT_READ
 import socket
 import time
 from typing import Optional
+from warnings import warn
 
 from jeepney import Parser, Message, MessageType, HeaderFields
-from jeepney.auth import SASLParser, make_auth_external, BEGIN, AuthenticationError
+from jeepney.auth import Authenticator, BEGIN
 from jeepney.bus import get_bus
+from jeepney.fds import FileDescriptor, fds_buf_size
 from jeepney.wrappers import ProxyBase, unwrap_msg
 from jeepney.routing import Router
 from jeepney.bus_messages import message_bus
 from .common import MessageFilters, FilterHandle, check_replyable
+
+__all__ = [
+    'open_dbus_connection',
+    'DBusConnection',
+    'Proxy',
+]
 
 
 class _Future:
@@ -46,27 +55,20 @@ def timeout_to_deadline(timeout):
 
 def deadline_to_timeout(deadline):
     if deadline is not None:
-        return deadline - time.monotonic()
+        return max(deadline - time.monotonic(), 0.)
     return None
 
 
-class DBusConnection:
-    def __init__(self, sock):
+class DBusConnectionBase:
+    """Connection machinery shared by this module and threading"""
+    def __init__(self, sock: socket.socket, enable_fds=False):
         self.sock = sock
+        self.enable_fds = enable_fds
         self.parser = Parser()
         self.outgoing_serial = count(start=1)
         self.selector = DefaultSelector()
         self.select_key = self.selector.register(sock, EVENT_READ)
-        self._unwrap_reply = False
-
-        # Message routing machinery
-        self.router = Router(_Future)  # Old interface, for backwards compat
-        self._filters = MessageFilters()
-
-        # Say Hello, get our unique name
-        self.bus_proxy = Proxy(message_bus, self)
-        hello_reply = self.bus_proxy.Hello()
-        self.unique_name = hello_reply[0]
+        self.unique_name = None
 
     def __enter__(self):
         return self
@@ -75,12 +77,81 @@ class DBusConnection:
         self.close()
         return False
 
-    def send(self, message: Message, serial=None):
-        """Serialise and send a :class:`~.Message` object"""
+    def _serialise(self, message: Message, serial) -> (bytes, Optional[array.array]):
         if serial is None:
             serial = next(self.outgoing_serial)
-        data = message.serialise(serial=serial)
-        self.sock.sendall(data)
+        fds = array.array('i') if self.enable_fds else None
+        data = message.serialise(serial=serial, fds=fds)
+        return data, fds
+
+    def _send_with_fds(self, data, fds):
+        bytes_sent = self.sock.sendmsg(
+            [data], [(socket.SOL_SOCKET, socket.SCM_RIGHTS, fds)]
+        )
+        # If sendmsg succeeds, I think ancillary data has been sent atomically?
+        # So now we just need to send any leftover normal data.
+        if bytes_sent < len(data):
+            self.sock.sendall(data[bytes_sent:])
+
+    def _receive(self, deadline):
+        while True:
+            msg = self.parser.get_next_message()
+            if msg is not None:
+                return msg
+
+            b, fds = self._read_some_data(timeout=deadline_to_timeout(deadline))
+            self.parser.add_data(b, fds=fds)
+
+    def _read_some_data(self, timeout=None):
+        for key, ev in self.selector.select(timeout):
+            if key == self.select_key:
+                if self.enable_fds:
+                    return self._read_with_fds()
+                else:
+                    return unwrap_read(self.sock.recv(4096)), []
+
+        raise TimeoutError
+
+    def _read_with_fds(self):
+        nbytes = self.parser.bytes_desired()
+        data, ancdata, flags, _ = self.sock.recvmsg(nbytes, fds_buf_size())
+        if flags & getattr(socket, 'MSG_CTRUNC', 0):
+            self.close()
+            raise RuntimeError("Unable to receive all file descriptors")
+        return unwrap_read(data), FileDescriptor.from_ancdata(ancdata)
+
+    def close(self):
+        """Close the connection"""
+        self.selector.close()
+        self.sock.close()
+
+
+class DBusConnection(DBusConnectionBase):
+    def __init__(self, sock: socket.socket, enable_fds=False):
+        super().__init__(sock, enable_fds)
+
+        # Message routing machinery
+        self._router = Router(_Future)  # Old interface, for backwards compat
+        self._filters = MessageFilters()
+
+        # Say Hello, get our unique name
+        self.bus_proxy = Proxy(message_bus, self)
+        hello_reply = self.bus_proxy.Hello()
+        self.unique_name = hello_reply[0]
+
+    @property
+    def router(self):
+        warn("conn.router is deprecated, see the docs for APIs to use instead.",
+             stacklevel=2)
+        return self._router
+
+    def send(self, message: Message, serial=None):
+        """Serialise and send a :class:`~.Message` object"""
+        data, fds = self._serialise(message, serial)
+        if fds:
+            self._send_with_fds(data, fds)
+        else:
+            self.sock.sendall(data)
 
     send_message = send  # Backwards compatibility
 
@@ -91,22 +162,7 @@ class DBusConnection:
         Otherwise, it will wait for up to timeout seconds, or indefinitely if
         timeout is None. If no message comes in time, it raises TimeoutError.
         """
-        deadline = timeout_to_deadline(timeout)
-
-        while True:
-            msg = self.parser.get_next_message()
-            if msg is not None:
-                return msg
-
-            b = self._read_some_data(timeout=deadline_to_timeout(deadline))
-            self.parser.add_data(b)
-
-    def _read_some_data(self, timeout=None):
-        for key, ev in self.selector.select(timeout):
-            if key == self.select_key:
-                return unwrap_read(self.sock.recv(4096))
-
-        raise TimeoutError
+        return self._receive(timeout_to_deadline(timeout))
 
     def recv_messages(self, *, timeout=None):
         """Receive one message and apply filters
@@ -114,7 +170,7 @@ class DBusConnection:
         See :meth:`filter`. Returns nothing.
         """
         msg = self.receive(timeout=timeout)
-        self.router.incoming(msg)
+        self._router.incoming(msg)
         for filter in self._filters.matches(msg):
             filter.queue.append(msg)
 
@@ -128,7 +184,10 @@ class DBusConnection:
         deadline = timeout_to_deadline(timeout)
 
         if unwrap is None:
-            unwrap = self._unwrap_reply
+            unwrap = False
+        else:
+            warn("Passing unwrap= to .send_and_get_reply() is deprecated and "
+                 "will break in a future version of Jeepney.", stacklevel=2)
 
         serial = next(self.outgoing_serial)
         self.send_message(message, serial=serial)
@@ -141,7 +200,7 @@ class DBusConnection:
                 return msg_in
 
             # Not the reply
-            self.router.incoming(msg_in)
+            self._router.incoming(msg_in)
             for filter in self._filters.matches(msg_in):
                 filter.queue.append(msg_in)
 
@@ -158,7 +217,9 @@ class DBusConnection:
         :param collections.deque queue: Matched messages will be added to this
         :param int bufsize: If no deque is passed in, create one with this size
         """
-        return FilterHandle(self._filters, rule, queue or deque(maxlen=bufsize))
+        if queue is None:
+            queue = deque(maxlen=bufsize)
+        return FilterHandle(self._filters, rule, queue)
 
     def recv_until_filtered(self, queue, *, timeout=None) -> Message:
         """Process incoming messages until one is filtered into queue
@@ -181,11 +242,6 @@ class DBusConnection:
             self.recv_messages(timeout=deadline_to_timeout(deadline))
         return queue.popleft()
 
-
-    def close(self):
-        """Close this connection"""
-        self.selector.close()
-        self.sock.close()
 
 class Proxy(ProxyBase):
     """A blocking proxy for calling D-Bus methods
@@ -220,9 +276,9 @@ class Proxy(ProxyBase):
             timeout = kwargs.pop('_timeout', self._timeout)
             msg = make_msg(*args, **kwargs)
             assert msg.header.message_type is MessageType.method_call
-            return self._connection.send_and_get_reply(
-                msg, timeout=timeout, unwrap=True
-            )
+            return unwrap_msg(self._connection.send_and_get_reply(
+                msg, timeout=timeout
+            ))
 
         return inner
 
@@ -238,22 +294,54 @@ def unwrap_read(b):
     return b
 
 
-def open_dbus_connection(bus='SESSION') -> DBusConnection:
-    """Connect to a D-Bus message bus"""
-    bus_addr = get_bus(bus)
+def prep_socket(addr, enable_fds=False, timeout=2.0) -> socket.socket:
+    """Create a socket and authenticate ready to send D-Bus messages"""
     sock = socket.socket(family=socket.AF_UNIX)
-    sock.connect(bus_addr)
-    sock.sendall(b'\0' + make_auth_external())
-    auth_parser = SASLParser()
-    while not auth_parser.authenticated:
-        auth_parser.feed(unwrap_read(sock.recv(1024)))
-        if auth_parser.error:
-            raise AuthenticationError(auth_parser.error)
 
-    sock.sendall(BEGIN)
+    # To impose the overall auth timeout, we'll update the timeout on the socket
+    # before each send/receive. This is ugly, but we can't use the socket for
+    # anything else until this has succeeded, so this should be safe.
+    deadline = timeout_to_deadline(timeout)
+    def with_sock_deadline(meth, *args):
+        sock.settimeout(deadline_to_timeout(deadline))
+        return meth(*args)
 
-    conn = DBusConnection(sock)
-    conn.parser.add_data(auth_parser.buffer)
+    try:
+        with_sock_deadline(sock.connect, addr)
+        authr = Authenticator(enable_fds=enable_fds)
+        for req_data in authr:
+            with_sock_deadline(sock.sendall, req_data)
+            authr.feed(unwrap_read(with_sock_deadline(sock.recv, 1024)))
+        with_sock_deadline(sock.sendall, BEGIN)
+    except socket.timeout as e:
+        sock.close()
+        raise TimeoutError(f"Did not authenticate in {timeout} seconds") from e
+    except:
+        sock.close()
+        raise
+
+    sock.settimeout(None)  # Put the socket back in blocking mode
+    return sock
+
+
+def open_dbus_connection(
+        bus='SESSION', enable_fds=False, auth_timeout=1.,
+) -> DBusConnection:
+    """Connect to a D-Bus message bus
+
+    Pass ``enable_fds=True`` to allow sending & receiving file descriptors.
+    An error will be raised if the bus does not allow this. For simplicity,
+    it's advisable to leave this disabled unless you need it.
+
+    D-Bus has an authentication step before sending or receiving messages.
+    This takes < 1 ms in normal operation, but there is a timeout so that client
+    code won't get stuck if the server doesn't reply. *auth_timeout* configures
+    this timeout in seconds.
+    """
+    bus_addr = get_bus(bus)
+    sock = prep_socket(bus_addr, enable_fds, timeout=auth_timeout)
+
+    conn = DBusConnection(sock, enable_fds)
     return conn